chore: restructure folder
