+++ packages/app/src/app/components/CreateSandbox/Icons/BackIcon.tsx
