+++ packages/app/src/app/components/CreateSandbox/Icons/CLIIcon.tsx
