+++ packages/app/src/app/components/CreateSandbox/Icons/DownloadIcon.tsx
