+++ packages/app/src/app/components/CreateSandbox/Icons/GitHubIcon.tsx
