+++ packages/app/src/app/components/CreateSandbox/Icons/Search.tsx
