+++ packages/app/src/app/components/CreateSandbox/Icons/TerminalIcon.tsx
