+++ packages/app/src/app/components/CreateSandbox/SearchBox/SearchBox.tsx
