+++ packages/app/src/app/components/CreateSandbox/availableTemplates.ts
