+++ packages/app/src/app/overmind/effects/api/index.ts
@@ -24,7 +24,7 @@ import {
   UserSandbox,
   SettingsSync,
 } from '@codesandbox/common/lib/types';
-import { LIST_PERSONAL_TEMPLATES } from 'app/components/CreateNewSandbox/queries';
+import { LIST_PERSONAL_TEMPLATES } from 'app/components/CreateSandbox/queries';
 import { client } from 'app/graphql/client';
 import { PendingUserType } from 'app/overmind/state';
